requirements.txt
# Core ML Framework
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision
opencv-python>=4.8.0
opencv-python-headless>=4.8.0
Pillow>=9.5.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Web Interface
streamlit>=1.25.0
streamlit-webrtc>=0.47.0

# Data Augmentation
albumentations>=1.3.0

# Utilities
PyYAML>=6.0
tqdm>=4.65.0
psutil>=5.9.0

# Optional: for better performance
# fastapi>=0.100.0
# uvicorn>=0.22.0